requirements.txt
google-genai>=1.0.0
markdown>=3.7
jinja2>=3.1.0
apscheduler>=3.10.0
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
